### 2018-06-30 #800 @whedon
[PRE REVIEW]: DeGAUSS: Decentralized Geomarker Assessment for Multi-Site Studies
**Submitting author:** @cole-brokamp (<a href="http://orcid.org/0000-0002-0289-3151">Cole Brokamp</a>)
**Repository:** <a href="https://github.com/cole-brokamp/DeGAUSS" target ="_blank">https://github.com/cole-brokamp/DeGAUSS</a>
**Version:** v0.2
