### 2018-06-30 4171f68
JOSS manuscript updates
