+++ paper.md
@@ -21,17 +21,13 @@ bibliography: paper.bib
 
 # Summary
 
-Health studies that utilize geocoding and assessment of any place-based characteristic, i.e. census-tract level income or estimated exposure to ambient air pollution frequently utilize residential addresses or other geolocation data that are considered protected health information (PHI).
+Health studies that utilize geocoding and assessment of any place-based characteristic, e.g. census-tract level income or estimated exposure to ambient air pollution, frequently utilize residential addresses or other geolocation data that are considered protected health information (PHI). The HIPAA privacy rule [@hipaa], the HITECH Act of 2009 [@hitech], and the Federal Policy for the Protection of Human Subjects [@commonrule] establish regulations to safeguard the confidentiality of patients and research subjects when health care providers or researchers use PHI. This is an outstanding challenge for multi-site studies because current approaches include getting institution-specific approval for a central site to conduct all analyses, which is a lengthy and sometime unfeasible process, or allowing each site to conduct their own analyses which requires expertise at each site and can result in non-reproducible and inconsistent geocoding and assessment of place-based characteristics, i.e. geomarkers. 
 
-The HIPAA privacy rule [@hipaa], the HITECH Act of 2009 [@hitech], and the Federal Policy for the Protection of Human Subjects [@commonrule] establish regulations to safeguard the confidentiality of patients and research subjects when health care providers or researchers use PHI.
+`DeGAUSS` is a standalone, container-based application that can produce geocodes and derive community and environmental exposures. Usable on PC, Mac, or Linux machines, identifying information never leaves the local machine. Figure 1 illustrates the process of using `DeGAUSS` within a multi-site study. Each study site uses `DeGAUSS` to both independently geocode their own addresses and link in the necessary place-based characteristics, strips any PHI, and then sends de-identified dataset out for analysis. In addition to securing PHI, this guarantees that the software will always run the same, regardless of its environment, which is a vital requirement for reproducible research. 
 
-This is an outstanding challenge for multi-site studies because current approaches include getting institution-specific approval for a central site to conduct all analyses, which is a lengthy and sometime unfeasible process, or allowing each site to conduct their own analyses which requires expertise at each site and can result in non-reproducible and inconsistent geocoding and geomarker assessment results. 
+![Illustration of DeGAUSS application within a multi-site study.](degauss_diagram.jpg)
 
-`DeGAUSS` is a standalone, container-based application that can produce geocodes and derive community and environmental exposures. Usable on PC, Mac, or Linux machines, identifying information never leaves the local machine. Each study site uses `DeGAUSS` to both independently geocode their own addresses and link in the necessary place-based characteristics, strips any PHI, and then sends de-identified dataset out for analysis. In addition to securing PHI, this guarantees that the software will always run the same, regardless of its environment, which is a vital requirement for reproducible research. The figure below illustrates the process of using `DeGAUSS` within a multi-site study.
-
-![DeGAUSS_diagram](degauss_diagram.jpg)
-
[…]
