+++ paper.bib
@@ -1,15 +1,18 @@
 @misc{hipaa,
 	author = "{United States Public Law}",
+  year = 1996,
 	title = {{Health Insurance Portability and Accountability Act of 1996 (HIPAA) Pub.L. 104–191 and the HIPAA Privacy Rule 2003. 45 CFR Part 160 and Part 16 Subparts A and E.}}
 }
 
 @misc{hitech,
 	author = "{United States Public Law}",
+  year = 2009,
 	title = {{Health Information Technology for Economic and Clinical Health (HITECH) Act of 2009. Section 13410(d).}}
 }
 
 @misc{commonrule,
 	author = "{United States Public Law}",
+  year = 1981,
 	title = {{Federal Policy for the Protection of Human Subjects (‘Common Rule’). 45 CFR part 46}}
 }
 
@@ -92,14 +95,13 @@
   }
 
 
-@article{@degauss-jamia,
+@article{degaussjamia,
 	Author = {Brokamp, Cole and Wolfe, Chris and Lingren, Todd and Harley, John and Ryan, Patrick},
 	Date-Added = {2018-02-23 20:52:28 +0000},
 	Date-Modified = {2018-02-23 20:53:58 +0000},
 	Doi = {10.1093/jamia/ocx128},
 	Issn = {1527-974X},
 	Journal = {Journal of the American Medical Informatics Association},
-	Keywords = {me},
 	Month = {March},
 	Number = {3},
 	Pages = {309--314},
