### 2018-06-30 #800 @cole-brokamp
Hi @whedon , I've updated the paper in the DeGAUSS repository based on the article proof above.  Please recompile.
