### 2018-06-30 #800 @cole-brokamp
Based on the list of existing JOSS reviewers, I feel that @nuest @jhollist @vsoch @sgrieve would all make great reviewers.
