### 2018-07-01 #800 @sgrieve
This looks like really interesting work! Unfortunately I'm not familiar enough with R to review this adequately.
