### 2018-07-01 #800 @vsoch
I am doing a trip followed by a cross country move starting next week, so I can't review imminently, but I'd be available at the end of July or early August!

I don't think R as a predominant language is totally necessary because the actual executable to review would be a container. For everyone in this thread, here is a direct link to the repo to help decide if you would be a suitable reviewer --> https://github.com/cole-brokamp/DeGAUSS. If you look at the README, interaction with the application is via docker commands.

@arfon is there a tag for container / Docker? I think many more that are familiar with containers (and perhaps not having the primary language as the executables inside) would still be able to provide some level of review. As long as one of the reviewers has more familiarity, I think it might be ok :L)
