### 2018-07-01 #800 @cole-brokamp
@vsoch makes a good point. In fact, this software is designed for users that don’t use R at all. Most of the end users hardly have any experience with command line programming at all.
