### 2018-07-01 #800 @arfon
> @arfon is there a tag for container / Docker? I think many more that are familiar with containers (and perhaps not having the primary language as the executables inside) would still be able to provide some level of review. As long as one of the reviewers has more familiarity, I think it might be ok :L)

We don't label submissions as such yet but I can definitely see the argument for doing so...
