### 2018-07-03 #800 @pjotrp
This is an interesting method for sure. I can be an editor if @vsoch is happy to review. There is no hurry.
