### 2018-07-03 #800 @pjotrp
@george-githinji are you Ok to review?
