### 2018-07-06 #800 @arfon
> @arfon hmm. So I would like @vsoch and @george-githinji as reviewers. Is that the case now?

Yep, looks to be. You can always refresh the page and look at the top of the issue thread:
