### 2018-07-08 #812 @pjotrp
@cole-brokamp we are starting review! To expedite matters do you mind going through above check list and make sure the reviewers can tick the boxes? You can't check them yourself, but you can make sure they are covered.
