### 2018-07-08 #812 @vsoch
here are updated review guidelines : https://joss.readthedocs.io/en/latest/review_criteria.html
