### 2018-07-08 #812 @cole-brokamp
Thanks, @pjotrp. I believe that the manuscript and repository checks all the boxes.  Looking forward to hearing from the reviewers.
