### 2018-07-09 cole-brokamp/DeGAUSS#10 @vsoch
[REVIEW] for Joss
Review
=======

This is software that can calculate coordinates based on addresses, provided in a container for reproducible use. It's well documented and described. My main comments have to do with clearing up some of the examples and description. I understand the use of containers for reproducibility, but it's not clear to me how geocoding an address (that can be reversed) somehow ensures privacy. Please read my comments below for details about how it might be better described (in human friendly terms for the layperson) and thank you for the opportunity to review this work. [Accept]

Criteria
===============

## Does the paper ONLY include these things?

- [x] A list of the authors of the software
  - confirmed that the main author is the only contributor to the repository
- [x] Author affiliations
- [x] A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience. _It may not be clear to some what geocoding is to begin with. There was a popular thing called "geocaching" which was hiding things at particular coordinates years back, and it might be confused with that (or generally other "geo" terms). A simple sentence like "Geocoding is the practice of..." would be sufficient for the layperson. It's also not clear to the layperson what a "container" is and how using one addresses some of the other issues. I can read this ok coming from a container + science background, but if not? I would have trouble._
- [x] A clear statement of need that illustrates the purpose of the software
   - _the statement is there (the first paragraph) but it is more "academically written" than "human friendly" and concise. You really just need to say "Geocoding might capture someone's private information, like where they live, and this is against HIPAA, etc.).  A shorter, more concise statement that gets the point across (with references to more verbose detail) would be more appropriate here. I can offer to help, it seems like what you want to say is:_

   1. Location-based data is important for different kinds of research
   2. It would go against rules to publish someone's personal location information
   3. We are missing standard ways to make this easy.

## General Review of Paper

 Here are a few things I genuinely don't understand / know what they are based on not being a part of the field:

  - "de-rive community and environmental exposures" (you mean like exposure to car exhaust? violence?)
  - "that can produce geocodes" (what does this mean it is doing? Does someone carry around a computer through a neighborhood and record things?)
 - "both indepen-dently geocode their own addresses and link in the necessary place-based characteristics" (I don't know what this means, maybe you would instead of stating this, walk the reader through an actual example?)
  - "sends de-identified dataset out for analysis" (how?)

Ah, here is where you start to explain it well:

> It was designed to be used by scientific researchers who wish to collect place-based data on study subjects and patients with a residential address.

The figure is nice, but a caption that more clearly explains it would be great!

- [x] Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it
- [x] A list of key references including a link to the software archive
- [x] confirmed that the paper doesn't have usage, API docs, etc.

## Review items

### Software

 - [x] Versions: _The tag on Github (v0.2) is different than the Docker image tags (v2.2). Could the author please comment on this discrepancy, and consider maintaining consist versions as to not be confusing for the user?_
__________________________________
Update September 10, 2018
The versions are coinciding with tags in a different repository, via an automated docker build. This is okay.
__________________________________

### Software license

 - [x] OSI approved, text file in respository.

Testing the software, I get these warning messages:

Loading required package: methods
  |============================================================| 100%, ETA 00:00
1: CBapply is deprecated but is maintained here for backwards compatability.
Consider using CB::mappp instead.
2: `as_function()` is deprecated; please use `as_mapper()` or `rlang::as_function()` instead
```
__________________________________
Update September 10, 2018
This warning message has been suppressed.
__________________________________

Given the container, has the author considered following the suggestion to eliminate the warning?

Here I am reviewing the README.md in the main repository.
 - [x] A statement of need
 - [x] The authors should clearly state what problems the software is designed to solve and who the target audience is. _Would the author consider moving the "About" to be closer to the top, so it's seen first?"
 - [x] There should be a clearly-stated list of dependencies. Ideally these should be handled with an automated package management solution. _Very well done, containers, and great documentation for installation, etc.!_
 - [x] Example usage. This isn't addressed directly, but would be easy to do! How about a list of bullet point examples for how this might be used? This would help me to understand better too :).
 - [x] The authors should include examples of how to use the software (ideally to solve real-world analysis problems). _The examples aren't provided, but the actual walk through of the example input file and using the container is spot on!_

__________________________________
Update September 10, 2018
The about section and example usage are fabulous.
__________________________________

One quick suggestion - the author provides a sample file with the column name `address` and then the example command is:

docker run --rm=TRUE -v "$PWD":/tmp degauss/geocoder my_address_file.csv addresses
```

I think it would make sense to 1) tell the user to name the file "my_address_file.csv" and then have the same column name (one or the other) so literally the user can copy paste the command (without change) from the $PWD with the file and it will work.

What isn't clear from this example is what I just did. I converted a file of addresses to... another file? Maybe describing this in the context of an actual example (e.g., "Let's say you have these addresses that represent... and you want to..." here is how you would do this.).

__________________________________
Update September 10, 2018
This file naming has also been updated and fixed.
__________________________________

 - [x] API documentation is good.
 - [x] For the next example, I don't know what geomarker assessment is (when you have latitute and longitude?) Perhaps each of these examples in the walk through would be best paired with a description of a real world example for noobs like me. ;) It might also make sense to continue the walkthrough from the previous step, first give the general instruction, and then show with the file that was just produced, e.g.:

docker run --rm=TRUE -v "$PWD":/tmp degauss/dist_to_major_roadway <name-of-geocoded-file>
docker run --rm=TRUE -v "$PWD":/tmp degauss/dist_to_major_roadway my_address_file_geocoded.csv
```

 - [x] Missing values - why would a value be missing, and is there some sort of partial information the viewer could return (in a separate file?) to investigate  or debug something?
 - [x] Question - if these two containers can reverse one another (e.g., I produce the geocoded file and then calculate a distance a major roadway) why are they maintained in separated containers? Has the author considered providing the software in a single container (to keep the versions consistent for the user - I won't download and use different ones) and then provide multiple entrypoints? Are the dependencies quite different?

#### References
There are a bunch of empty lines in the references (maybe authors?) that might be looked at.

#### Community guidelines
 - [x] Community guidelines are missing. The author has a clear link to getting help, but could you please add a CONTRIBUTING.md with instructions for how to contribute?
 - [x] Functionally, I tested the examples in the README and it works as expected. Is the author intending for more advanced users to ever shell into the image and interact directly with the software?
 - [x] There are no tests. Has the author considered setting up continuous integration?
