### 2018-07-09 #812 @vsoch
I have finished my review, reported in issue here: https://github.com/cole-brokamp/DeGAUSS/issues/10 I will go through the points in this repository and confirm when the issues have been addressed!
