### 2018-07-09 cole-brokamp/DeGAUSS#10 @vsoch
And by the way, this is really cool! Given standard inputs and outputs, a lot of cool visualization tools could be built around this!
