### 2018-07-18 cole-brokamp/DeGAUSS#10 @cole-brokamp
Hi @vsoch, I just wanted to stop by and say thank you for the valuable feedback and review.  I just moved to a new house and am up against a couple of academic deadlines :weary:, but I intend on updating the software according to your input as soon as I can.  Thanks again!
