### 2018-07-18 cole-brokamp/DeGAUSS#10 @vsoch
Congratulations on your move!! No worries, please take the time you need. I'm in the middle of moving too (the truck is still... who knows where...) and I can completely empathize. I'm mostly avoiding thinking about the "I need to unpack when things finally arrive" part _crawls into fetal position and thinks of someday-when-I-am-not-moving_
