### 2018-08-08 #812 @cole-brokamp
OK, great.  Thanks for your patience — grant season is almost finished and then I’ll be there!

> On Aug 8, 2018, at 10:27 AM, Pjotr Prins <notifications@github.com> wrote:
>
> No deadlines - just interest :)
>
