### 2018-09-10 fd2722a
JOSS revisions (addresses #10)
