### 2018-09-10 cole-brokamp/DeGAUSS#10 @cole-brokamp
I've added a commit that has addressed all of your comments, @vsoch.  Please review and let me know what you think when you have time.  Thanks again for the helpful and constructive feedback!  I'm happy to see `DeGAUSS` improving as a result of the peer review process.  I've outlined my changes below, roughly in order of your review above:

# Changes Made for JOSS Review

## Added summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience

- defined "geocoding"
- defined "container" and how it addresses some of the scientific issues

## Added clear statement of need that illustrates the purpose of the software

- made need statement more "human friendly" and concise

## Add clarifications to paper

- defined and added examples of community and environmental exposures
- removed the confusing term "produce geocodes"
- referred the reader to the README for an example, including how to share deidentified data
- added better caption to the figure

## Sofware Changes

- The GitHub tags do indeed match the DockerHub tags -- its just that `DeGAUSS/geocoder` is hosted in a separate [GitHub repo](https://github.com/cole-brokamp/geocoder). This provides separation between "geocoding" and "geomarker assessment". The matched tags are achieved through automated builds on DockerHub triggered by a new tagged release on GitHub.
- suppressed warning messages for usage of `CB::CBapply`

## Documentation

- Moved "About" section to top so it is seen first
- Made sure that the problems that the software can solve are more clearly stated and identified who the target audience is
- Added a list of specific example usages in the "About" section
- Adjusted code example so user can copy/paste if using file named `my_address_file.csv` with `address` column
- Added output of example usage and extended it to geomarker assessment
- Added why some geocodes might be missing and how users can troubleshoot this problem.
- I have chosen to keep `DeGAUSS` tools in separate containers because their dependencies are quite different and `DeGAUSS` is designed to be extendable by the user for other geomarker datasources and assessment toolkits. The geospatial datasets used are commonly > 1 GB and keeping each geomarker in a separate container not only reduces the image size, but more importantly, keeps the entire system modular and more reproducible.

## References

- The references missing authors are referring to legislation, which is commonly cited without an author given the understanding that it is "United States Public Law"

## Community Guidelines

- Add community guidelines with a `CONTRIBUTING.md`
- I do not expect advanced users to shell into the image and interact directly with the software.  DeGAUSS is designed specifically for users that want to geocode and do geomarker assessment, but do not have the technical expertise to use the underlying software.  If a user is familiar and capable with these processes, then the only advantage of DeGAUSS would be its reproducibility -- it would likely prevent more of an obstacle than an assistance. Although I have left this out in the README, I often use this technique to troubleshoot with others remotely and will add it to the wiki page.
