### 2018-09-10 #812 @cole-brokamp
I've revised the manuscript, code, and README based on @vsoch's review.  You can see a summary of my changes here: https://github.com/cole-brokamp/DeGAUSS/issues/10#issuecomment-419918139
