### 2018-09-10 cole-brokamp/DeGAUSS#10 @vsoch
Awesome! Will have some time to look later today, stay tuned!
