### 2018-09-10 cole-brokamp/DeGAUSS#10 @vsoch
# Review after changes
Hey @cole-brokamp ! Apologies for the delay, my state (NC) is preparing for a storm of doom and everyone is freaking out, myself included because I evacuated my apartment this morning after the predicted storm category increased and debating for a bit. Now I'm at least somewhere with a generator, and will be assured internet connectivity! I can proceed with the review!

 - The README.md and abstract are so much better! As someone not familiar with the domain you teach me everything that is important to know about the software - both from a rationale for "Why are you doing this period for science" to "How is this tool important for that goal?" The description of containers is also just at the right level. You can only take on so much education for every little piece of technology, and it's up to the reader to do some google searching in the case of wanting more information.
 - I see the matching versions now, thank you for pointing me to that. +1 on the automated builds! I like this feature too.
 - All of my needs are addressed here, and great job with the community guidelines and contributing docs.

You have responded very well to all of my points, and I would suggest this paper for publication. I will go back to the original thread to look over the paper once more.
