### 2018-09-10 #812 @vsoch
Bravo @cole-brokamp! You have done a beautiful job fixing up the paper and repository. I have two quick checks to do before signing off, both rather trivial.

- tests: do you describe anywhere how to test? Are there automated tests of any kind? The general goal would be that if a developer stumbles across any of the code and wants to quickly check "is this functioning as it should on my system" you can point them to the things to run.
 - paper rending: I noticed there seems to be a missing / not rendered double slash before "a process"

![image](https://user-images.githubusercontent.com/814322/45324890-77007d80-b51d-11e8-8339-821931eecb66.png)

You might not have any control over this, but it might be worth glancing over. Those are the only two remaining (small) things I would want to ask about before signing off.
