### 2018-09-13 cole-brokamp/DeGAUSS#10 @cole-brokamp
Thank you for the quick followup and I hope that you are staying safe!!  I had one question about implementing tests.  This is something that I want to do, but am struggling with the specifics.  Do you think it would be wise to implement some formal unit tests?  Or are you thinking more along the lines of "Here is a sample address file and if you run it through these three containers using this specific code, then here is what the output should look like at each step"?  Does that make sense? I appreciate any insight you might have.
