### 2018-09-13 cole-brokamp/DeGAUSS#10 @vsoch
hey @cole-brokamp! You would minimally want those functions run against the data files on each commit to master, using a service like Travis CI or Circle CI. It’s something you don’t have to think about, and any user can run the equivalent tests manually. I can offer to create a simple setup for you (and issue a PR, you would need to officially connect the service from your web interface) to help you learn by just seeing what I mean. What do you think?
