### 2018-09-13 cole-brokamp/DeGAUSS#10 @cole-brokamp
I see.  I am familiar with tests in R, but have no experience creating test for bash commands.  I don't want to create any additional work for you -- can you point me in a helpful direction and I can ask you for more help if I need it?
