### 2018-09-13 cole-brokamp/DeGAUSS#10 @vsoch
Sure! Here you go https://circleci.com/docs/2.0/getting-started/  what you can maximally do is set up a build test and deploy cycle where the containers will be built and deployed to Docker Hub after passing the tests. But if you prefer to just push manually this is overkill, and you can just have a simple setup that builds the containers and runs the same commands the user would. You basically are going to write a .circleci/config.yml file that will issue these commands to test! Yaml and their workflows are a bit funky so please ask me if you need help. I’ve made these so many times I can do it pretty quickly. Good luck!
