### 2018-09-13 cole-brokamp/DeGAUSS#10 @cole-brokamp
thanks!  I will dive in this weekend and let you know how it goes
