### 2018-09-16 #812 @cole-brokamp
I’m working on adding some automated build testing to address the last comment. I’ll let you know when it’s ready for a final review.

> On Sep 16, 2018, at 1:25 PM, Pjotr Prins <notifications@github.com> wrote:
>
> @cole-brokamp ping!
>
