### 2018-09-16 #812 @vsoch
+1! We had good discussion here https://github.com/cole-brokamp/DeGAUSS/issues/10#issuecomment-421098449 about tests and @cole-brokamp is tackling this probably as we speak! And I’m ready at bay for when Cole needs me!
