### 2018-10-02 c8545b9
add crude CI testing
