+++ .travis.yml
@@ -1,0 +1,17 @@
+sudo: required
+
+services:
+  - docker
+
+before_install:
+  - docker build -t degauss/acs_income ACS_income
+  - docker build -t degauss/census_tracts census_tracts
+  - docker build -t degauss/dist_to_major_roadway dist_to_major_roadway
+
+script:
+  - docker run --rm=TRUE -v "$PWD":/tmp degauss/acs_income my_address_file_geocoded.csv
+  - test -e my_address_file_geocoded_medianhouseholdincome.csv || exit 1
+  - docker run --rm=TRUE -v "$PWD":/tmp degauss/census_tracts my_address_file_geocoded.csv
+  - test -e my_address_file_geocoded_censustracts.csv || exit 1
+  - docker run --rm=TRUE -v "$PWD":/tmp degauss/dist_to_major_roadway my_address_file_geocoded.csv
+  - test -e my_address_file_geocoded_disttomajorroadway.csv || exit 1
