### 2018-10-02 2179d6d
upgrade to latest version of gdal in r_geo base image
