+++ ACS_income/deps.yaml
@@ -6,7 +6,7 @@
   Version: 0.3.3
 - Package: rgdal
   Repository: CRAN
-  Version: 1.2-5
+  Version: 1.3-3
 - Package: dplyr
   Repository: CRAN
   Version: 0.5.0
