### 2018-10-02 e1c2d1d
small type in JOSS paper
