### 2018-10-02 cf5ba47
add spiffy travis-ci badge :sunglasses:
