### 2018-10-02 cole-brokamp/DeGAUSS#10 @cole-brokamp
I (finally!) got time to add tests through Travis.  Each of the three images is built and a command is ran on each using a sample geocoded address file and Travis checks to make sure the expected output file is present.  Thanks for your help in teaching me about this!
