### 2018-10-02 #812 @cole-brokamp
I've addressed the last comment by implementing tests (see https://github.com/cole-brokamp/DeGAUSS/issues/10#issuecomment-426444872 for more details).  What do you think, @vsoch?
