### 2018-10-02 cole-brokamp/DeGAUSS#10 @vsoch
Fantastic work, and your tests just pass! This definitely meets all the criteria for publication, and after the fact when you feel up for it, you can give a go at adding a deploy step so in the future the docker hub builds are just magically done for you (tagged with either a software version or Github commit). I'll update the main thread to let the editors know! I'm going to close here.
