### 2018-10-02 #812 @vsoch
@cole-brokamp has done a fantastic job adding tests (and we are green and good to go!) and so I've updated my review and am very happy to say that @cole-brokamp has addressed all of my concerns, and I my recommendation is accept for publication in Joss. Great job @cole-brokamp! Passing forward to you again @pjotrp.

https://travis-ci.org/cole-brokamp/DeGAUSS
