### 2018-10-03 3b0d179
bump to v0.3 for JOSS publication
