### 2018-10-03 #812 @cole-brokamp
Most of my references are either software vignettes or United States Public Law, both of which do not have DOIs available.  I've included all other DOIs that are available for my references.

I have tagged a release as v0.3 that includes all changes made during the JOSS review.  I have an automated build setup on Zenodo for all releases, so the current DOI will point to the latest version: 10.5281/zenodo.570873

I'm waiting on the Zenodo automated archiving for v0.3, which will be the release linked to the JOSS publication -- I'll update here when the DOI pointing specifically to v0.3 is available.
