### 2018-10-03 #812 @cole-brokamp
Here is the DOI for the v0.3 release:

10.5281/zenodo.1443521

Thanks again
