### 2018-10-04 #812 @cole-brokamp
Thanks everyone for their help in getting this manuscript published.  It has been a great experience and I look forward to reviewing for JOSS soon.
